requirements.txt
requests>=2.28.1
openai>=1.0.0
google-api-python-client>=2.70.0
python-dotenv>=1.0.0
aiohttp>=3.8.3
async-timeout>=4.0.2
tenacity>=8.1.0
pytest>=7.0.0